integrate hecstac with existing codebase
