+++ pyproject.toml
@@ -36,13 +36,16 @@ dependencies = [
     "mypy_boto3_s3==1.34.120",
     "numpy==2.0.1",
     "psutil==6.0.0",
-    "pystac==1.10.0",
+    "pystac==1.13.0",
     "pyarrow==16.1.0",
     "rasterio==1.3.10",
     "requests==2.32.3",
     "shapely==2.0.5",
     "xarray==2024.11.0",
-    "rioxarray==0.18.1"
+    "rioxarray==0.18.1",
+    "rashdf==0.8.2",
+    "typing_extensions==4.14.1",
+    "obstore==0.6.0"
 ]
 
 [project.optional-dependencies]
