+++ ripple1d/api/app.py
@@ -12,7 +12,8 @@ from werkzeug.exceptions import BadRequest
 
 from ripple1d.api import tasks
 from ripple1d.api.utils import get_unexpected_and_missing_args
-from ripple1d.ops.fim_lib import create_fim_lib, create_rating_curves_db, fim_lib_stac, nwm_reach_model_stac
+from ripple1d.hecstac.ras_to_gpkg import gpkg_from_ras
+from ripple1d.ops.fim_lib import create_fim_lib, create_rating_curves_db
 from ripple1d.ops.metrics import compute_conflation_metrics
 from ripple1d.ops.ras_conflate import conflate_model
 from ripple1d.ops.ras_run import (
@@ -22,7 +23,6 @@ from ripple1d.ops.ras_run import (
 )
 from ripple1d.ops.ras_terrain import create_ras_terrain
 from ripple1d.ops.subset_gpkg import extract_submodel
-from ripple1d.ras_to_gpkg import gpkg_from_ras
 
 app = Flask(__name__)
