+++ ripple1d/ops/endpoints.py
@@ -7,7 +7,8 @@ import sys
 from typing import Callable
 
 from ripple1d.consts import SUPPRESS_LOGS
-from ripple1d.ops.fim_lib import create_fim_lib, create_rating_curves_db, fim_lib_stac, nwm_reach_model_stac
+from ripple1d.hecstac.ras_to_gpkg import gpkg_from_ras
+from ripple1d.ops.fim_lib import create_fim_lib, create_rating_curves_db
 from ripple1d.ops.metrics import compute_conflation_metrics
 from ripple1d.ops.ras_conflate import conflate_model
 from ripple1d.ops.ras_run import (
@@ -17,7 +18,6 @@ from ripple1d.ops.ras_run import (
 )
 from ripple1d.ops.ras_terrain import create_ras_terrain
 from ripple1d.ops.subset_gpkg import extract_submodel
-from ripple1d.ras_to_gpkg import gpkg_from_ras
 from ripple1d.ripple1d_logger import RippleLogFormatter  # , initialize_process_logger
 
 func_lookup = {
@@ -30,14 +30,13 @@ func_lookup = {
     "run_incremental_normal_depth": run_incremental_normal_depth,
     "run_known_wse": run_known_wse,
     "create_fim_lib": create_fim_lib,
-    "nwm_reach_model_stac": nwm_reach_model_stac,
-    "fim_lib_stac": fim_lib_stac,
     "create_rating_curves_db": create_rating_curves_db,
 }
 
 
 def execute_endpoint(func: str, args: dict):
     """Execute the specified function."""
+    print("at api lookup")
     return func_lookup[func](**json.loads(args))
